ESLint start
Started working on ESLint rules. Harder than I thought and a lot of
stuff that needs to be checked.

I've edited the code (from master), to work with the rules I've
implemented so far. A noteable exception is that the rules "no-undef"
and "no-unused-vars" shows some warnings. I'm not quite certain about
the inclusion of these rules yet but I'll include them in this commit
anyway.