Prop and default bear picture and colorPicker v1
Added props so that a bear picture can be choosen sending in a value as
a prop to bear. Also added some colors to bears to send as props to
bear.

Tried to create the colorPicker. Right now, it's always on but hopefully
by the next update it'll only pop up for a bear when it's clicked.