Delete bear action
Added a deleteBear action to redux. Have not yet implemented it for use
in react however (I've put it where it belongs, though). For now, it
does not care how many bears are currently being used.

I also renamed all instances of 'UpdateBearArray' and it's variations to
'UpdateBear' since it isn't really an array.