Back to game button
Added a button to the startview that moves the player back to their last
game. The button will not be visible if:
* The player completed the last game and found all combinations
* The player don't have a recorded last game

Game will be 'saved' for this button when the player clicks the
startView button in any view. Saved data will be removed as soon as the
player goes back to the last game.